requirements.txt
fastapi>=0.115.6
uvicorn>=0.34.0
anthropic>=0.42.0
sqlalchemy>=2.0.36
python-dotenv>=1.0.1
pydantic>=2.10.3
